Simplified the App.
I bit off a little more than I could chew with a tutorial I was following.

It created a CRUD folder that had generic paths for communicating with the DB.

While I understood what was going on, couldn't really get it to work.

Simplified back down for the purposes of this assignment.